refactor(Movie): 뷰에서 사용하는 MovieUiModel 분리
